requirements.txt
wandb
tqdm==4.66.1
ema_pytorch==0.2.3
accelerate==0.23.0
transformers==4.34.0
pytorch_fid==0.3.0
matplotlib==3.8.0
pynvml==11.5.0
tensorboard
git+https://github.com/2toinf/DecisionNCE.git
git+https://github.com/Farama-Foundation/d4rl@master#egg=d4rl